about cmd: make ApplicationCommand-only
